* Backport compiler fixes from 5.1   * Fix warnings with Win64.   * Fix IP/PGM Router Alert with Win64.   * Add further restrict usage.   * Remove unused pthread_rwlock_t from Windows slim reader/writer lock usage.   * Update SNMP & HTTP modules for new peer/socket usage.   * Add error checking for non-Win32, non-GNU compilers in thread module.   * Fix select usage in recovery of hard failure on network sends on POSIX platforms. * Bump micro version.

+++ openpgm/pgm/include/pgm/skbuff.h
@@ -174,7 +174,7 @@ pgm_skb_headroom (
 	const struct pgm_sk_buff_t*const skb
 	)
 {
-	return (char*)skb->data - (char*)skb->head;
+	return (uint16_t)((char*)skb->data - (char*)skb->head);
 }
 
 static inline
@@ -183,7 +183,7 @@ pgm_skb_tailroom (
 	const struct pgm_sk_buff_t*const skb
 	)
 {
-	return (char*)skb->end - (char*)skb->tail;
+	return (uint16_t)((char*)skb->end - (char*)skb->tail);
 }
 
 /* reserve space to add data */
